#23 move connectors to extension folder
