+++ src/extension/background-script/connectors/lnbits.js
