+++ src/extension/background-script/connectors/lnd.js
