+++ src/extension/background-script/connectors/native.js
