+++ src/extension/background-script/index.js
@@ -3,7 +3,7 @@ import browser from "webextension-polyfill";
 import utils from "../../common/lib/utils";
 import Settings from "../../common/lib/settings";
 import Accounts from "../../common/lib/accounts";
-import connectors from "../../common/lib/connectors";
+import connectors from "./connectors";
 
 import initLsatInterceptor from "./lsatInterceptor";
